feat: add POW to email sending routes
The registration and password reset routes now require a proof of work
by making sure the SHA512 hash of the request body begins with 16 zero
bits. These are the two routes that could be abused for sending emails
to arbitraty addresses. The nonce only needs to be calculated once per
email address but emails are only sent once every 30 minutes anyway.